수정사항 반영

+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/cart/service/CartService.java
@@ -224,17 +224,16 @@ public class CartService {
      */
     @Transactional
     public Long deleteCartItem(Long cartItemId, Long memberId) {
-        Member member = memberRepository.findById(memberId).orElseThrow(
-                () -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
 
-        Cart cart = cartRepository.findById(memberId).orElseThrow(
+        Cart cart = cartRepository.findByMemberId(memberId).orElseThrow(
                 () -> new CartException(CartErrorCode.CART_NOT_FOUND));
-
         CartItem cartItem = cartItemRepository.findById(cartItemId).orElseThrow(
                 () -> new CartException(CartErrorCode.CART_ITEM_NOT_FOUND));
+        if (!cartItem.getCart().getId().equals(cart.getId())) {
+            throw new CartException(CartErrorCode.CART_INVALID);
+        }
 
         cartItemRepository.deleteById(cartItemId);
-
         return cartItemId;
     }
 }
